Layoffs_Analysis_2020_2023/README.md
This folder contains the Layoffs Analysis 2020-2023 project, which explores trends in layoffs across various industries over the years using SQL. The analysis includes data cleaning, data transformation, and exploratory analysis focusing on trends, funding impact, and industry volatility. All SQL scripts used for the analysis are available in this folder, along with the documentation to reproduce and understand the findings.